lib/tools: provide functions to (un)parse sets of integers
In POSIX/Linux API there are various objects that may be interpreted
as sets of integer numbers, including various bitmasks, file
descriptor sets or CPU id sets. Such objects regularly need to be
converted to and from a string representation which is identical
in all cases. So a set of functions is provided to do such a conversion
in a generic way. Further patches would provide instantiations for
file descriptor and CPU sets.

OL-Review-Id: 12226
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>