lib/confapi: add type-safe funcs to get instances versions
Since there are many of integer types for every integer type there
should be a separate type-safe function of cfg_get_instance_fmt()
and cfg_get_instance_sync_fmt() for values of such type. The version for
int type in reality is working with int32_t type and it is only for
backward compatibility.

OL-Redmine-Id: 11890
Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>