lib/tad: size packet mmap Rx ring using HW descriptor count
A test may want to receive large number of packets. Even if
it sets larger HW descriptor count (in example, 4096) using
Configurator, the packet mmap Rx ring will still default to
MAX 1024 entries. That appears to be insufficient sometimes.

Prefer sizing the mmap ring by the current descriptor count
of the NIC. If that fails, fall back to the legacy approach.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>