lib/tools: get rid of a timer's signal
When a timer expires and generates a signal, that signal can affect
some other routines as well. For instance, it makes read_socket() fail
with EINTR.

Actually it is not mandatory to use asynchronous notifification with
signal events when the timer expires bacause it is still not available
to the end user who has to use a certain API to check the timer status.

OL-Redmine-Id: 11566
Breaks: now a timer cannot be deleted when it expires;
        use explicit te_timer_stop() to release it
Signed-off-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Ivan Malov <ivan.malov@oktetlabs.ru>