lib/tapi_fio: use integer values for size spec
Using string values for sizes in TAPI fio was apparently
done to support passing values with prefixes as test parameters.
Now we have scaled-value test parameters, and it is much more
convenient if the size is an integral value, if it is used in
some calculations.

OL-Redmine-Id: 12503
Breaks: tapi_fio (struct tapi_fio_opts)
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>