engine/configurator: improve searching for an instance
When an instance is added locally (to be committed to TA later),
cfg_db_find() adds it direct children to database when looking for
them if they are missing. It allows to set them locally too.
However it does not work for more remote descendants of a locally
added instance, making it hard to construct and commit to TA a
complex object. This patch fixes it.

OL-Redmine-Id: 12545
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>