doc/cm: correct config files to qualify the schema
Despite the fact that the schema for YAML isn't written the changed
lines stand out from the general. And they should be changed if we will
use new code that works with cyaml.

The `depends:` value should always be a sequence. And `oid` shouldn't
be omitted.

Despite `access` is optional in the parser code now there is no reason
to think that default value `read_create` is the correct default value.
And I think that `access` should always present. Here is the only case
that I've found where access is omitted in configuration.

OL-Redmine-Id: 11888
Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>