lib/tools: implement simple JSON serializing functions
te_mi_logger have a lot of ad hoc JSON formatting which
is hard to maintain and is not robust. This patch provides
a consistent API to generate JSON values, which is going
to be used by te_mi_logger later. The API can also be used
to support JSON generation in tapi_tool for tools that use
JSON config format.

The API is not intended to replace a full-fledges JSON library,
its primary usecase is serialization of relatively simple C
structures where using e.g. Jansson would lead to a lot of
unnecesary boilerplate.

OL-Redmine-Id: 12514
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>