doc: add an option to check doc quickly
Running the full generation cycle takes quite a long
even without the global namespace, and most apparent
documentation errors are reported by Doxygen. In addition,
the list of warnings produced by Doxygen is a bit messy.
This patch provides a quick way to see all Doxygen-related
problems of the code base in a clean manner.

OL-Redmine-Id: 9521
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>