doc/cm: reflect PCI serial number dependency on driver
Prevent Configurator from trying to restore PCI serial numbers of
devices that were rebound to a different driver during test execution.

PCI serial numbers are extracted using the devlink interface, which
is not available with every driver. For example, when vfio-pci is used,
a blank serial number is reported by the Configurator.

By restoring the driver first, the Configurator reverses the change
to the serial number instance implicitly, thus avoiding the need to
restore it explicitly.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>