lib/tools: add functions for random buffers with constraints
In many cases it is desired to generate a random sequence subject
to certain constraints, e.g. being a sequence of digits or suitable
for use as a filename etc. The patch extends older buffer-making
API with such functions. Older functions are retained as special-case
instantiations of more generic functions.

OL-Redmine-Id: 12392
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Tested-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>