lib/tapi_tool: report to MI logger truncated wrk arguments
The approach of increasing the size of http request via
appending a query string to the requested url naturally inflates
the size of wrk tool command line. We report the command line into MI
logger thus we get in the tests logs the vast output. To make MI logs
neat, we report to the logger truncated wrk arguments each of which
does not exceed 256 characters.

OL-Redmine-Id: 12664
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>