engine/configurator: add libcyaml dependency
The XML structures was used for temporary storing parsed
information from YAML configuration files. There was a decision
to eliminate XML from CS since there are some features that
doesn't provided by XML but only by YAML.

The fundamental idea behind CYAML is to allow applications to construct
schemas which describe both the permissible structure of the YAML
documents to read/write, and the C data structure(s) in which the loaded
data is arranged in memory. Nevertheless that CYAML couldn't handle
configuration files (they are not so simple as CYAML expected) it is
very convenient to use it for backup and for dynamic history files.

Breaks: libcyaml library is needed on engine
Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@arknetworks.am>