dispatcher: add env exporting support
Make it possible to export environment variables not only via script
files but also via option ones.

Before it was the only way to set options and environments all together
in a single file: via '--script', e.g.
  export DUT_TYPE=arm
  TE_EXTRA_OPTS="${TE_EXTRA_OPTS} --trc-tag=arm"
  TE_EXTRA_OPTS="${TE_EXTRA_OPTS} --tester-req=!NO_ARM"

You may notice TE_EXTRA_OPTS is overused in order to achieve that.
Also impossibility to set all related settings as options may lead to
spreading them through several files and hence breaking consistency.

Suggested solution is designed to get rid of these disadvantages.
Just look how laconic it could be if it passed via '--opts':
  --export=DUT_TYPE=arm
  --trc-tag=arm
  --tester-req=!NO_ARM

Note it is disabled to export value free variables with this way
because there is no a practical use of them.

OL-Redmine-Id: 11416
Signed-off-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>