doc/cm: add a dependency for ftpserver OID in the schema
When the configuration files from doc/cm are included in TS, the
tests start to fail with the CS-ENOENT error in the main prologue.
This is caused by the lack of dependency for the ftpserver object.
In the previous TS configuration files, before I removed redundant
object registrations from there, this dependency was there.
To avoid errors when working with the configuration in TS, need to
add а dependency to the attributes of the ftpserver object.

OL-Redmine-Id: 12813
Fixes: afaada7087af ("Import the first public release")
Signed-off-by: Nikolay Klepikov <nikolay.klepikov@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>