apps: implement application to print logs about core dumps
The new application will set location for core files and then wait for
new core files. On discovering every new core file it will check whether
it was generated for one of the binaries in TA directory, and if it is
the case, it will print a log about it including backtrace from gdb.
This will simplify debugging segfaults in TA and RPC server.

OL-Redmine-Id: 12092
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>