lib/rcfunix: improve stdin redirection for commands
Currently stdin is replaced with /dev/null for commands run in
lib/rcfunix - but it is done via ssh option "-n". It means that it works
only for remote TAs but not for local ones, and also makes it hard to
run a command and communicate with it via its standard input (because
"-n" is hardcoded in ssh command prefix).
So it seems better to avoid "-n" ssh option and instead pass
TE_EXEC_CHILD_DEV_NULL_FD value to te_shell_cmd() for currently used
commands. It should have the same effect as "-n" option, and it is
easy to pass pointer to file decriptor instead when writing something
to standard input is required.

OL-Redmine-Id: 12092
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Daniil Byshenko <daniil.byshenko@oktetlabs.ru>