lib/rpcc_dpdk: do not duplicate RPC input arguments
On an RPC call, input arguments are only read to create an RPC external
representation. Despite it, some calls create copies of data to pass
them as input arguments of RPC calls.
Of course, the copying overhead is negligible. But the RPC wrapper are
expected to be pretty short and simple.

Thus, make the text of RPC wrappers slightly simpler.

OL-Redmine-Id: 12750
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>