doc/cm: correct the access of L2TP-related OID in the schema
When the configuration files from doc/cm are included in TS, the TS
started returning an error for all l2tp tests. This is caused by the
Configurator trying to create objects with the "read_write" access
mode. These objects are actually not "read_write", but "read_create".
To avoid errors when working with the configuration in TS, need to
change the access mode with "read_create" for certain OIDs.

OL-Redmine-Id: 12850
Fixes: afaada7087af ("Import the first public release")
Signed-off-by: Nikolay Klepikov <nikolay.klepikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>