tools/rgt: improve line numeration in HTML logs
Do not restart log lines numeration in every new log table in the same
file. Multiple log tables may be present in log of package or session
where references to child packages/sessions/tests are shown between
such tables. Line numbers should be unique in a HTML log file because
they are used in HTML links to specific places in the log.

OL-Redmine-Id: 12861
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>