lib/tools: use dynamic strings for variable expansion
Old API is retained for compatibility, but the new code
should use new functions:
- te_expand_parameters => te_string_expand_parameters
- te_expand_env_vars => te_string_expand_env_vars
- te_expand_kvpairs => te_string_expand_kvpairs

Save for the output argument, the new functions provide
the same functionality as the old ones. In particular,
the limitation on the number of positional arguments
(TE_EXPAND_MAX_POS_ARGS, i.e. 10) is not new, it has only
been embedded into the function signatures.

OL-Redmine-Id: 12845
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>