engine/tester: emit argument hash for all iterations
When a test is run as part of a prologue, it does not get a TIN, and
therefore does not emit its argument hash. This prevents some tools
from telling different iterations of this test apart from each other.

Fix this problem by attaching the argument hash whenever it's possible.

Link: https://github.com/oktetlabs/test-environment/pull/14
Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Acked-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>