lib/tapi_env: allow binding addresses while reusing ports
When a server may be hidden behind NAT, a test is likely to need to deal
with two different addresses: one will be used by the server to bind to
the port, and another will be used by the client to reach the server.

TEST_GET_ADDR_REUSE_PORT provides a way for a test to store both
addresses without having to manipulate the ports manually. With it,
either of the addresses is first taken using the old TEST_GET_ADDR
macro, then it is passed to the new macro along with the name of the
other address.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Ivan Malov <ivan.malov@arknetworks.am>