agents/unix: implement OvS flow handling
Tests need a way to set up and modify OvS flow rules, to redirect
or otherwise process traffic.

Processing OvS flow rules isn't trivial: different OvS tools use
slightly differing formats. In order to deal with these discrepancies
and to simplify the implementation in the test agent, most of the logic
was moved to a new library called ovs_flow. In the future, this library
can be extended to allow more complex processing of flow rules.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Ivan Malov <ivan.malov@arknetworks.am>