doc/cm: expose switchdev information
This information can be used by tests to identify and control switchdev
entities, e.g. representors.

A separate object provides a mapping from switch ID and port name to the
name of the interface with these properties. NIC drivers usually use
PF's switch or port ID as representor switch ID and some vendor-specific
human-readable name as representor port name. This object allows tests
to find the network interface associated with a representor, and then
grab it as a resource.

Technically, this object depends on the drivers bound to NICs and their
SR-IOV configuration. However, the actual instances of this object are
not available for some time after a NIC is bound to another driver.
Because of this, configuration rollback may fail. To avoid these timing
issues, either a small delay needs to be inserted after some operations
related to driver-related changes, or this object needs to be excluded
from the rollback process. Achieve the latter by declaring the object
volatile.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Ivan Malov <ivan.malov@arknetworks.am>