lib/tapi: add test param type for expected results
Some tests may have parameters specifying the expected
outcome of a certain action, which may be either succcess
with a given result or a failure with a given error code.
There was no support in TE for mapping te_errno codes in
test parameters, and the usecase above would require two
aligned parameters or some custom non-trivial syntax.

This patch adds a special kind of test parameters that
allow to specify the expected status code and the expected
result in one value. While mostly aimed at unit-like tests,
it may find its use in production test suites as well.

OL-Redmine-Id: 12949
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>