lib/tools: add a function for syntactic pathname joining
We already have functions to construct filenames from parts,
but they assume local filesystem and check for the file existence.
Yet it is in many cases useful to construct a file on the Engine
side (e.g. from a test) to be used on the agent side. This patch
implements such a function that is purely syntactic.

OL-Redmine-Id: 12954
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>