lib/tools: allow null result for conversion functions
Passing NULL as a pointer to the result to conversion
functions is pretty useful, as it allows to use such
functions as validity checkers without creating dumb
variables. Previously, handling of NULL was different
for different conversion functions: some accepted it,
some returned an error. The patch unifies the behaviour.

OL-Redmine-Id: 12971
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Daniil Byshenko <daniil.byshenko@oktetlabs.ru>