lib/tools: add macro to map strings to actions
A task of dispatching an action depending on a string value
is pretty common. It is possible to do it with pre-existing
string-to-enum mapping functions and then using an enum value
as an index to an array of function pointers. However, such an
enum would in many cases be totally ad hoc, so a dedicated
dispatch macro is implemented for this purpose.

OL-Redmine-Id: 13011
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>