lib/rpcs_dpdk: fix creation of an IPv4 item with zero value
If the field has a null value in the ASN representation, it is
treated as "there is no value". However, from the user's point
of view, the zero value is still a value and the user expects
DPDK to create an item to match by the zero value. The creation
of an item should be rejected only if the mask is zero.

Fixes: afaada7087af ("Import the first public release")
Signed-off-by: Artemii Morozov <artemii.morozov@arknetworks.am>
Reviewed-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>