lib/tools: add function to process lines in a buffer
The function allows to call a handler for each complete
line in a buffer (i.e. such that it is terminated by
a new line character) and keep the rest for the following
processing, thus providing an easy and robust way to
handle a stream of bytes as a sequence of lines.

OL-Redmine-Id: 13043
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>