tools/rgt: use signed integer for nesting level
rgt-conv uses signed integer for representing nesting level. And it
actually can write negative nesting level in XML log, for instance
when log is split into multiple pages inside log bundle and the first
message in the current page is not at the top of depth hierarchy.
rgt-xml2json should handle it, not crash with assertion.

Fixes: 894b7825909d ("tools/rgt: implement a tool for getting JSON logs")
OL-Redmine-Id: 12861
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>