scripts: make less strict checks for work in progress
Developers may want to check patches which are not yet
finalized, and in such case some checks make little sense,
such as the presence of Reviewed-by trailers. This patch
makes it possible to disable such checks either by an explicit
option to te_check_patches or when pushing to a user private
branch.

Fixes: e99c4cb30852 ("scripts: provide Git hook to block invalid commits")
Fixes: d9ba98c3882e ("scripts: add checks for extra trailers in git hook")
Fixes: 5375e2c70b05 ("scripts/hooks: allow to configure production branches")
Fixes: 33d9e970bc89 ("scripts: enhance remote branch detection for patch checker")
OL-Redmine-Id: 13095
Suggested-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>