lib/tapi_bpf: improve adding buffers to FILL ring
Replace socket pointer with UMEM/interface/queue_id in arguments
of rpc_xsk_rx_fill_simple() to allow to add buffers to FILL ring
before a socket is created. On some NICs (for instance, Intel E810)
driver expects that buffers are already added when AF_XDP socket is
created and prints warning if it is not the case.

OL-Redmine-Id: 13159
Breaks: rpc_xsk_rx_fill_simple()
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>