engine/tester: allow to adjust dial weights in XML
The new attribute will allow to set a coefficient to be applied
to default selection weight of a given run item when choosing
iterations randomly with help of --dial option. With help
of this attribute some tests or test packages can be excluded
completely, others can be selected less or more often than by
default.

OL-Redmine-Id: 13160
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>