lib/tapi: replace system allocators with TE equivalents
Functions from TAPI Mem now use TE allocators instead
of system functions for consistency. Technically there is
a slight change in behaviour, as they now abort on memory
allocation failures instead of failing the test; however,
this is extremely unlikely to happen in practice. In future,
fatal errors in tests would be treated as test failures.

OL-Redmine-Id: 13239
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Anatoly Mirin <anatoly.mirin@oktetlabs.ru>