lib/tapi: fix ISO C violation
ISO C (unlike C++) does not permit const variables to
be used as initializers for static variables. However,
GCC started to accept it even in maximum compliance
mode from version 8.3. So the code that compiles fine
on newer distros start to fail on older distros, but
what's more important is that Clang does not accept
this code as well.

Fixes: d4495ffcc3ba ("lib/tapi_tool: implement TAPI for trex")
OL-Redmine-Id: 13222
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>