lib/tools: allow reading files in chunks
One of the initial goals of te_file_read_string
was reading files in /proc filesystem on the agent.
However, it turns out it could not do that, because
the size of such files is always reported as zero.
The patch amends the function to be able to read
such files in chunks.

Fixes: 19eab56b47a2 ("lib/tools: add API to read/write from/into dynamic strings")
OL-Redmine-Id: 13264
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>