doc/cm: mark eswitch objects volatile
The values of these objects are very much driver-specific and can change
at almost any point. For example, with some drivers, changing the MAC
address of a given interface can lead to a change in its port ID.

This driver-specific nature makes it hard to predict all factors that
affect the values of these objects. Therefore, to avoid guessing, these
objects can be considered volatile.

Fixes: adede836a18c ("doc/cm: expose switchdev information")
Link: https://github.com/oktetlabs/test-environment/pull/32
Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Artem Andreev <artem.andreev@oktetlabs.ru>