lib/ta_restconf: add /restconf/config/search CS tree
RESTCONF server does not allow to get a list of
root subtrees, so, we need to know which subtrees to try.
Also some inner or upper subtrees can be lost
because of server bugs.

Proposed usecase is a preconfigured array in cs.conf.x file
that checked in /restconf/data CS tree list method
to include the subtrees that really exist.

Runtime creatable objects is out of scope so far,
possible it will be handled while /restconf/data
CS subtree implementation.

OL-Redmine-Id: 13472
Signed-off-by: Michael Kochkin <michael.kochkin@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>