agent/unix: decrease log levels of some errors in CPU config
Some systems do not support some of the processor attributes whose
values the configurator tries to retrieve. Failure to retrieve these
values is considered an error, but it is not. Therefore, it is necessary
to lower the logging level to a warning and use special loggers to print
log messages once.

Fixes: b301c4bba829 ("agents/unix: add cache type to CS database")
OL-Redmine-Id: 12619
Signed-off-by: Alexander Dubovoj <alexander.dubovoj@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>