agents/unix: avoid breaking interface synchronization
Check whether private flags can be obtained before listing them
in configuration tree. It is possible that list of private flags
can be retrieved but their values cannot. For instance, this was
observed for Solarflare Communications SFC9220 10/40G Ethernet
Controller. In such case do not list private flags in configuration
tree to avoid breaking configuration synchronization for an interface.

OL-Redmine-Id: 13583
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>