lib/tapi_tool: fix typos when using variables in TRex
The cross-use of the Rx and Tx values is most likely a typo, since it
is not commented on. In addition, it makes sense to store these values
the way TRex outputs them. If any test wants to swap the Rx and Tx,
then it should take care of it itself.

Fixes: d4495ffcc3ba ("lib/tapi_tool: implement TAPI for trex")
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>