lib/loggerta: try to avoid truncating logged messages
The log message can be truncated if the message size exceeds
the LOGFORK_MAXLEN value. Increasing this value solves this
for known use cases - wrk-tool and TRex.

This patch is an attempt to fix the current problem and does
not guarantee that messages will never be truncated.
The right solution should be implemented in the future.

OL-Redmine-Id: 12075
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>