tools/rgt: use engineering notation for graph axis labels
Very large and very small quantities are very hard to read
in fixed point decimal notation, so the patch makes the plotting
library show values larger than 1e5 or smaller than 1e-5 in
the engineering notation (which is the exponential notation with
exponents always being multiple of 3). A self test is added as
a demo of rendering graph with such values.

OL-Redmine-Id: 13669
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>