engine/tester: rework test spawning to avoid shell
We do not need shell to run a test.  Moreover, using shell implies
that all parameters should be escaped and that it is not easy to send
a signal to the test (because we know the pid of shell and do not know
the pid of the test).  Remove shell from this place.

OL-Redmine-Id: 13751
Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>