common: use standard ISO C boolean type
The boolean type is part of ISO C since C99 and
it is going to be part of the core language in C23,
so there is absolutely no sense in keeping a legacy
custom boolean type. The patch removes all uses of
te_bool and corresponding constants and mark its
definition as deprecated. No visible behaviour may
have changed, since the type was already defined as
an alias to ISO C bool.

OL-Redmine-Id: 13768
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>