lib: allow libraries to be linked wholly
In some cases, especially when it comes to external libraries,
it's necessary to link the entire library to make sure the ELF
constructors work reliably.

One potential use case is TS-provided Configurator subtrees.
A test suite can provide a library that implements a set of
Configurator objects and registers itself in the ELF constructor.

Link: https://github.com/oktetlabs/test-environment/pull/45
Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>