lib: allow libraries to be linked wholly
In some cases, especially when it comes to external libraries, it's necessary
to link the entire library to make sure the ELF constructors work reliably.

One potential use case is TS-provided Configurator subtrees. A test suite can
provide a library that implements a set of Configurator objects and registers
itself in the ELF constructor.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>