lib/tapi_tool: process optional TRex filters separately
A TRex report can contain a large number of counters for both
the client and server side, and adding new optional filters in
the same way as the mandatory filters will be too cumbersome
and inconvenient.
Now, the tapi_trex_opt_flt_cl_get() and tapi_trex_opt_flt_srv_get()
functions are designed to get the value of the optional filter.

Breaks: TS using tcps_sndbyte and tcps_rcvbyte values in reports
OL-Redmine-Id: 13796
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>