lib/tapi: use te_test_name to check that test is monitor
Use te_test_name to check that test without te_test_id is a monitor.
Test based command monitor uses TE_LOG_ID_UNDEFINED to add logs from this
monitor inside tests running in parallel. But the default behaviour of
TEST_START is to stop the test for this ID.

But any command monitor has a special name ("tester_monitorX" where X is the
ordinal number of monitor). We can use this name to avoid EXIT_FAILURE for
monitors.

Signed-off-by: Oleg Sadakov <Oleg.Sadakov@arknetworks.am>