engine/tester: implement ability to start test based command monitor
Tester starts this monitor with TE_LOG_ID_UNDEFINED as te_test_id. This is
useful to merge log from this monitor inside tests running in parallel.

But the default behaviour of TEST_START is to stop the test for this ID.
So you have to update your test to take it into account.

At the end of related session tester will send SIGUSR1 signal to this monitor.

Signed-off-by: Oleg Sadakov <Oleg.Sadakov@arknetworks.am>