agents/unix: specify dependencies for RCF listener explicitly
Separate the dependency list from the one used for the main TA executable.
Otherwise it's possible for one of the TA dependencies to export an ELF
constructor that would bring with it a dependency on a library that requires
its user to implement some set of functions.

Of course, the build fails if this happens.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>