agents/unix: specify deps for RCF listener explicitly
Separate the dependency list from the one used for the main TA
executable. Otherwise it's possible for one of the TA dependencies
ti export an ELF constructor that would bring with it a dependency on
a library that requires its user to implement some set of functions.

Of course, the build fails if this happens.

Link: https://github.com/oktetlabs/test-environment/pull/45
Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>