lib/tapi_rdma_perf: use corresponding name for stats entry
Since we will use statistics with several entries it is reasonable to
have two different additional types for the whole statistics and for
one entry only. Also we rename stats to stats_entry in cases where only
one entry is used.
By now the types for statistics coincides with the type for statistic
entry but later we will change the type for statistics.

Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@arknetworks.am>
Signed-off-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>