lib/tools: provide a function to compare iovecs
The existing buffer comparison function is only able
to compare several copies of an expected content
against a single actual buffer. Yet it is quite common
that the expected content is built from several chunks
and the actual content is scattered across several
buffers. So this patch provides a function to compare
two sequences of bytes scattered across buffers in any
fashion.

OL-Redmine-Id: 14068
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>