engine/tester: strip indentation of the multi-line values
Strip indentation of the multi-line values if they're starting with
'new line' character.

This is required in case of:

      <value>
      first
          sub-line
      second
      </value>

That turns into

char * = ""
         "      first"
         "          sub-line"
         "      second"

While in reality we need:

char * = "first"
         "    sub-line"
         "second"

Keep internal indention and remove the external/inherited from XML.

Note, that this commit will likely break history of all tests in all
systems if they have multi-line variables. This is unlucky, but I don't
see how this can be avoided unless we add tester conf option or
something.

Signed-off-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>