engine/tester: process XML text nodes in a more correct way
Tester config parsing code relied on a semi-documented feature
of libxml2 that text and comment nodes have specific pseudonames.
That also means that the parser would also accept real <text>
element nodes where a text node is expected. The patch changes
the parser to use the official libxml2 way of testing for text
and comment nodes.

OL-Redmine-Id: 14028
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>