lib/tad: always compute L4 datagram size from IP packet size
CSAP supports 'advanced' checksum matching that can actually
validate the checksum instead of comparing against a defined
fixed value in the match pattern. To do that, the code needs
to know the datagram size, which is currently taken from the
datagram buffer size, and, in the case of the whole Ethernet
frame size being less than minimum frame size, it's adjusted.

However, should the FCS/CRC be present in the frame for some
reason, the calculation goes haywire and yields wrong values.

Take advantage of the preceding L3 match, where one can take
L3 payload size from the header and stash that for later use.

Link: https://github.com/oktetlabs/test-environment/pull/62
Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Reviewed-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>