agents/unix: fix accidentally overlooked substring API use
The API for substrings has been changed in v1.45.0:
the return value of most functions is now either a boolean
success flag or a counter, so the sign of success in both
cases is a non-negative value (which would be treated as
error before). Most uses of these function in TE have
already been fixed, but one particuliar use has slipped
the attention of the original patch's author.
So this place is fixed too now.

GitHub-TE-Id: 63
OL-Redmine-Id: 14058
Fixes: ebc5b2fd184a ("lib/tools: sanitize substrings API")
Reported-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>