lib/loggerta: add function for closing logger socket
There are some cases, where we need to start RPC server
from the same application many times with different TE runs.
It means, that logger will have different ports to connect.
So we need a way to close logger socket, to create new one to work with
new TE run.

Link: https://github.com/oktetlabs/test-environment/pull/90
Signed-off-by: Artemii Morozov <artemii.morozov@arknetworks.am>
Acked-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>