lib/tools: provide a high-quality RNG
The function rand_range() that is used extensively in
TE has some drawbacks:
- since it is based on C standard rand(), on a typical
  Linux system it only produces 31-bit values, and using
  several consecutive rand() calls to produce a wider
  value may result in unfortunate bit correlation;
- rand() is not the best RNG in itself and the way
  rand_range() is implemented on top of it may make
  lower bits of the result significantly less random.

This patch introduces a fair 64-bit PRNG that is based
on a recent algorithm called xoshiro256++ and some utility
functions around it. It also provide a self test that
ensures the PRNG passes the statistical test for uniformity.

Link: http://vigna.di.unimi.it/ftp/papers/ScrambledLinear.pdf
OL-Redmine-Id: 14268
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>